feat(CustomerInfo): Set up appointment with barberId
