+++ src/app/App.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { CustomerInfo, AllBarbersContainer } from '../components';
+import { CustomerInfoContainer, AllBarbersContainer } from '../components';
 import './App.css';
 
 // TODO: Relocate the CustomerInfo Component later on;
@@ -7,7 +7,7 @@ class App extends Component {
   render() {
     return (
       <div className="App">
-        <CustomerInfo />
+        <CustomerInfoContainer />
         <AllBarbersContainer />
       </div>
     );
