+++ src/components/index.js
@@ -1,3 +1,3 @@
 // Barrel file for components, which will be used in to compose our main App component
-export { default as CustomerInfo } from './CustomerInfo';
+export { default as CustomerInfoContainer } from './containers/CustomerInfo';
 export { default as AllBarbersContainer } from './containers/AllBarbers';
